KVM: PPC: MPIC: Expose in-kernel pic capabilities
On x86, one would call kvm_irqchip_create() to initialize an in-kernel
interrupt controller. That function then goes ahead and initializes
global capability variables as well as the default irq routing table.

On ppc, we can't call kvm_irqchip_create() because we can have different
types of interrupt controllers. So we want to do all the things that
function would do for us in the in-kernel device init handler.

Signed-off-by: Alexander Graf <agraf@suse.de>