migration/postcopy: not necessary to discard all RAM at the beginning
ram_discard_range() unmap page for specific range. To be specific, this
clears related page table entries so that userfault would be triggered.
But this step is not necessary at the very beginning.

ram_postcopy_incoming_init() is called when destination gets ADVISE
command. ADVISE command is sent when migration thread just starts, which
implies destination is not running yet. This means no page fault
happened and memory region's page tables entries are empty.

This patch removes the discard at the beginning.

Signed-off-by: Wei Yang <richardw.yang@linux.intel.com>